refactor: 서비스 명령과 쿼리 작업 분리하기
