+++ app/src/test/java/com/codesoom/assignment/application/ProductCommandServiceTest.java
@@ -1,0 +1,143 @@
+package com.codesoom.assignment.application;
+
+import com.codesoom.assignment.domain.Product;
+import com.codesoom.assignment.exception.ProductNotFoundException;
+import com.codesoom.assignment.repository.ProductRepository;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+@DataJpaTest
+@DisplayName("ProductCommandService 클래스")
+class ProductCommandServiceTest {
+
+    private final String name = "My toy";
+    private final String maker = "toy shop";
+    private final int price = 15000;
+    private final String imageUrl = "toy.jpg";
+
+    @Autowired
+    private ProductRepository productRepository;
+
+    private ProductCommandService productCommandService;
+
+    @BeforeEach
+    void setUp() {
+        productCommandService = new ProductCommandService(productRepository);
+    }
+
+    @AfterEach
+    void clear() {
+        productRepository.deleteAll();
+    }
+
+    private Product getProduct() {
+        return new Product(name, maker, price, imageUrl);
+    }
+
+    @Nested
+    @DisplayName("create 메소드는")
+    class Describe_create {
+
+        @Test
+        @DisplayName("상품을 생성해 반환한다")
+        void it_returns_new_product() {
+            Product createdProduct = productCommandService.create(getProduct());
+
+            assertThat(createdProduct).isNotNull();
+            assertThat(createdProduct.getId()).isNotNull();
+        }
+    }
+
+    @Nested
+    @DisplayName("update 메소드는")
+    class Describe_update {
+
+        @Nested
+        @DisplayName("수정이 가능한 경우")
+        class Context_with_valid_id {
+            private Long id;
+            private Product source;
+
+            @BeforeEach
+            void setUp() {
+                Product product = productCommandService.create(getProduct());
+                id = product.getId();
+
+                final String updatePrefix = "updated_";
+                final int updatedPrice = 20000;
+
+                source = new Product(updatePrefix + name,
+                        updatePrefix + maker,
+                        updatedPrice,
+                        updatePrefix + imageUrl);
+            }
+
+            @Test
+            @DisplayName("수정된 상품을 반환한다")
+            void it_returns_updated_task() {
+                Product updatedProduct = productCommandService.update(id, source);
+
+                assertThat(updatedProduct).isNotNull();
+                assertThat(updatedProduct.getId()).isEqualTo(id);
+                assertThat(updatedProduct.getName()).isEqualTo(source.getName());
+            }
+        }
+
+        @Nested
+        @DisplayName("수정이 불가능한 경우")
+        class Context_with_invalid_id {
+            private final Long invalidId = 1000L;
+
+            @Test
+            @DisplayName("예외를 던진다")
+            void it_throws_exception() {
+                assertThatThrownBy(() -> productCommandService.update(invalidId, getProduct()))
+                        .isInstanceOf(ProductNotFoundException.class);
+            }
+        }
+    }
+
+    @Nested
+    @DisplayName("delete 메소드는")
+    class Describe_delete {
+
+        @Nested
+        @DisplayName("삭제 가능한 id가 주어지면")
+        class Context_with_valid_id {
+            private Long id;
+
+            @BeforeEach
+            void setUp() {
+                Product product = productCommandService.create(getProduct());
+                id = product.getId();
+            }
+
+            @Test
+            @DisplayName("삭제를 수행한다")
+            void it_removes_product() {
+                productCommandService.deleteById(id);
+            }
+        }
+
+        @Nested
+        @DisplayName("삭제 불가능한 id가 주어지면")
+        class Context_with_invalid_id {
+            private final Long invalidId = 1000L;
+
+            @Test
+            @DisplayName("예외를 던진다")
+            void it_throws_exception() {
+                assertThatThrownBy(() -> productCommandService.deleteById(invalidId))
+                        .isInstanceOf(ProductNotFoundException.class);
+            }
+        }
+    }
+}
